distributed/pyproject.toml
[project]
name = "distributed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.9",
    "google-adk>=1.16.0",
    "uvicorn>=0.35.0",
]
